add transaksi

+++ app/Http/Controllers/Admin/KategoriController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Products;
 use Illuminate\Http\Request;
 use RealRashid\SweetAlert\Facades\Alert;
 use Illuminate\Support\Facades\Validator;
@@ -15,7 +16,6 @@ class KategoriController extends Controller
     {
         $this->middleware("auth");
     }
-
 
     public function index()
     {
@@ -35,7 +35,6 @@ class KategoriController extends Controller
         return view('admin.form.form_kategori', $data);
     }
 
-
     public function store(Request $request)
     {
         $validator = Validator::make($request->all(), [
@@ -46,13 +45,22 @@ class KategoriController extends Controller
         ]);
 
         if ($validator->fails()) {
-            toast($request->aksi.' gagal dilakukan','danger');
+            toast($request->aksi.' gagal dilakukan','error');
             return redirect()->back()->withErrors($validator)->withInput();
         }
+
+        $produk     = Products::where('kategori_produk', '=', $request->kategori_produk)->get();
 
         $posts = Kategori::updateOrCreate(['id' => $request['id']], [
             'kategori_produk' => $request->kategori_produk,
         ]);
+
+        if ($produk)
+        {
+            $kategoriProduk = $request->kategori_produk;
+
+            Products::query()->update(['kategori_produk' => $kategoriProduk]);
+        }
 
         if($posts)
         {
@@ -66,7 +74,6 @@ class KategoriController extends Controller
         }
     }
 
-
     public function edit($id)
     {
         $data = [
@@ -79,14 +86,21 @@ class KategoriController extends Controller
 
     public function destroy($id)
     {
-        $kategori = Kategori::find($id);
+        $kategori   = Kategori::find($id);
+        $produk     = Products::first();
 
         if (!$kategori) {
-            abort(404); // Tambahkan ini untuk menangani kasus kategori tidak ditemukan
+            abort(404);
         }
 
-        $kategori->delete();
-        toast('Hapus kategori berhasil dilakukan!', 'success');
+        if ($produk->kategori_produk === $kategori->kategori_produk)
+        {
+            toast('Hapus kategori gagal dilakukan karena terdapat produk terkait!', 'error');
+        }else{
+            $kategori->delete();
+            toast('Hapus kategori berhasil dilakukan!', 'success');
+        }
+
 
         return redirect()->route('kategori.index'); // Gantilah dengan rute yang sesuai
     }
